refactor(utils): restructure react related utils
