+++ packages/lumx-react/src/components/list/useInteractiveList.tsx
@@ -2,8 +2,8 @@ import { ListItemProps } from '@lumx/react';
 import { isClickable } from '@lumx/react/components/list/ListItem';
 
 import { isComponent } from '@lumx/react/utils/type';
-import { flattenChildren } from '@lumx/react/utils/flattenChildren';
-import { mergeRefs } from '@lumx/react/utils/mergeRefs';
+import { flattenChildren } from '@lumx/react/utils/react/flattenChildren';
+import { mergeRefs } from '@lumx/react/utils/react/mergeRefs';
 import get from 'lodash/get';
 import {
     cloneElement,
