+++ packages/lumx-react/src/components/notification/Notification.tsx
@@ -10,7 +10,7 @@ import { NOTIFICATION_CONFIGURATION } from '@lumx/react/components/notification/
 import { GenericProps, HasTheme } from '@lumx/react/utils/type';
 import { getRootClassName, handleBasicClasses } from '@lumx/react/utils/className';
 import { useTransitionVisibility } from '@lumx/react/hooks/useTransitionVisibility';
-import { mergeRefs } from '@lumx/react/utils/mergeRefs';
+import { mergeRefs } from '@lumx/react/utils/react/mergeRefs';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
