+++ packages/lumx-react/src/components/popover/useRestoreFocusOnClose.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { getFirstAndLastFocusable } from '@lumx/react/utils/browser/focus/getFirstAndLastFocusable';
-import { OnBeforeUnmount } from '@lumx/react/utils/OnBeforeUnmount';
+import { OnBeforeUnmount } from '@lumx/react/utils/react/OnBeforeUnmount';
 import type { PopoverProps } from './Popover';
 
 /**
