+++ packages/lumx-react/src/components/progress-tracker/ProgressTracker.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 
 import { GenericProps } from '@lumx/react/utils/type';
 import { getRootClassName } from '@lumx/react/utils/className';
-import { mergeRefs } from '@lumx/react/utils/mergeRefs';
+import { mergeRefs } from '@lumx/react/utils/react/mergeRefs';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
 import { useRovingTabIndex } from '../../hooks/useRovingTabIndex';
